templates-service/requirements.txt
fastapi
pydantic_settings
uvicorn
pymongo
python-docx
httpx
requests